[refactor] 임장이미지뷰컨 이미지 더하기, 삭제하기 온보딩 로직 추가

+++ juinjang/Projects/App/Sources/UserInterface/RecordingRoom/ImjangImageList/ImjangImageListViewController.swift
@@ -174,8 +174,16 @@ final class ImjangImageListViewController: BaseViewController {
                 case .popButtonTap:
                     self.popView()
                 case .trashButtonTap:
+                    if UserDefaultManager.shared.isOnboarding {
+                        self.present(SignUpBottomSheetView(), animated: true)
+                        return
+                    }
                     self.deleteImageButtonTapped()
                 case .addButtonTap:
+                    if UserDefaultManager.shared.isOnboarding {
+                        self.present(SignUpBottomSheetView(), animated: true)
+                        return
+                    }
                     self.addImage()
                 default: break
                 }
@@ -328,7 +336,8 @@ extension ImjangImageListViewController: UICollectionViewDelegate, UICollectionV
 
 // MARK: ImagePicker Delegate
 extension ImjangImageListViewController: PHPickerViewControllerDelegate {
-    func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
+    func picker(_ picker: PHPickerViewController,
+                didFinishPicking results: [PHPickerResult]) {
         picker.dismiss(animated: true)
         let group = DispatchGroup()
